Update networkaccessmanager.cpp (#13195)
QT emits updates as well as start and end when a recource is received. Currently these are ignored and the number of bytes returned (bodySize) is incorrect as it is just set to the size of the first read which is returned by start. The fix adds a new option 'data' which returns the bytes (bodySize) returned in subsequent reads. If when OnResourceRecevieved is triggered for a particular resource the bytes (bodySize) from 'start' and subsequent 'read' status are added together the correct size of the resource will be calculated. 'end' does not return bodysize. This does not rely on content-length being passed by the server for instance when chunking is being used, or when gzip compression is being used as QT specifically removes the Content-Length header and does not return it to phantomjs